requirements.txt
# Core ML and Data Science packages
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web frameworks
flask>=2.2.0
fastapi>=0.95.0
uvicorn>=0.20.0
flask-cors>=4.0.0

# Data and caching
redis>=4.3.0

# Core utilities
pydantic>=1.10.0
python-dotenv>=0.19.0
requests>=2.28.0
aiohttp>=3.8.0

# Monitoring and system
prometheus-client>=0.15.0
psutil>=5.9.0
schedule>=1.1.0

# Configuration
pyyaml>=6.0
click>=8.1.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=4.0.0
requests>=2.28.0
aiohttp>=3.8.0
azure-ml-client>=1.0.0
azure-storage-blob>=12.14.0
azure-keyvault-secrets>=4.6.0
azure-identity>=1.12.0
prometheus-client>=0.15.0
psutil>=5.9.0
schedule>=1.1.0
pyyaml>=6.0
click>=8.1.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
